Surveillance System/requirements.txt
opencv-python
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
python-dotenv
pyserial
requests
